# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course API",
        "description": "API documentation for authentication, courses, interests, favorites, students, and teachers.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/v1/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List all courses",
                "operationId": "3522f88b734e2061541a8dbd0b6be53c",
                "responses": {
                    "200": {
                        "description": "Courses list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Laravel API",
                                        "description": "Build REST APIs with Laravel.",
                                        "teacher_id": 2,
                                        "interest_id": 1,
                                        "price": 300
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a course",
                "operationId": "92ce73d19fae55890bc7e00c723b7557",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel API"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Build REST APIs with Laravel.",
                                        "nullable": true
                                    },
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "interest_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 300
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "Laravel API",
                                "description": "Build REST APIs with Laravel.",
                                "teacher_id": 2,
                                "interest_id": 1,
                                "price": 300
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course created successfully.",
                                    "course": {
                                        "id": 1,
                                        "name": "Laravel API",
                                        "description": "Build REST APIs with Laravel.",
                                        "teacher_id": 2,
                                        "interest_id": 1,
                                        "price": 300
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "The name field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/courses/favorites": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get favorite courses for authenticated student",
                "operationId": "cd6aed30fa0f0a7a72b6743947402ed4",
                "responses": {
                    "200": {
                        "description": "Favorite courses list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 3,
                                        "name": "PHP OOP",
                                        "description": null,
                                        "teacher_id": 5,
                                        "interest_id": 2,
                                        "price": 0
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Only students can access favorite courses."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/courses/matching-interests": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get courses matching authenticated student interests",
                "operationId": "2463f51369f1f118a88adff6d9f2e8c3",
                "responses": {
                    "200": {
                        "description": "Matched courses list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 7,
                                        "name": "Python Data",
                                        "description": "Data analysis basics.",
                                        "teacher_id": 4,
                                        "interest_id": 2,
                                        "price": 200
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Only students can access courses by interests."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Show one course",
                "operationId": "e041494670480ff24d76c7191143774c",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 1,
                                    "name": "Laravel API",
                                    "description": "Build REST APIs with Laravel.",
                                    "teacher_id": 2,
                                    "interest_id": 1,
                                    "price": 300
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "No query results for model [App\\Models\\Course] 999"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update a course",
                "operationId": "c57e5081c7a5e3a381be7b32611032f6",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel API Advanced"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Advanced API topics.",
                                        "nullable": true
                                    },
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "interest_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 400
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "Laravel API Advanced",
                                "description": "Advanced API topics.",
                                "price": 400
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course updated successfully.",
                                    "course": {
                                        "id": 1,
                                        "name": "Laravel API Advanced",
                                        "description": "Advanced API topics.",
                                        "teacher_id": 2,
                                        "interest_id": 1,
                                        "price": 400
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "price": [
                                            "The price must be an integer."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "No query results for model [App\\Models\\Course] 999"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "operationId": "abe203e12027b72fdc6ba6c593413aa4",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course deleted successfully."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "No query results for model [App\\Models\\Course] 999"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}